requirements.txt
groq
chromadb
sentence-transformers
fastapi
uvicorn
python-dotenv
pypdf